requirements.txt
apiflask==2.0.2
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
psycopg2-binary==2.9.9

